Add stubs for sending and handling packets
Add functions for sending different types of packets used in the AODV
protocol, and empty functions that are called when those packets are
received.

We need to be able to send 24 bytes in a packet for the RREQs, so there
is not enough space to just use a `radio.sendBuffer(...)`. This means
that we need to be able to send "raw" packets over the radio which, as
far as I could tell, can only be done in C++ with a call to
`uBit.radio.datagram.send`.

The `radioEnable` function is taken directly from the existing radio
module. We will need to think about how we will handle setting the
group based on which party we are joining.

Only one function can be registered as a callback on radio events at
once, so we will likely have to disable the radio module from being
used at the same time as our module.

Buffers are used to communicate between TypeScript and C++ because a
maximum of 4 `number` or `string` arguments are supported in a C++
function when compiled in this way (Attempting to use more throws an
assertion error on compilation).